requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
openai>=1.0.0

